+++ hal/armv7a/zynq7000/interrupts.c
@@ -184,7 +184,7 @@ int hal_interruptsSetHandler(intr_handler_t *h)
 }
 
 
-char *hal_interruptsFeatures(char *features, unsigned int len)
+char *hal_interruptsFeatures(char *features, size_t len)
 {
 	(void)hal_strncpy(features, "Using GIC interrupt controller", len);
 	features[len - 1U] = '\0';
