+++ hal/armv7r/zynqmp/timer.c
@@ -140,7 +140,7 @@ static void hal_timerSetPrescaler(u32 freq)
 }
 
 
-char *hal_timerFeatures(char *features, unsigned int len)
+char *hal_timerFeatures(char *features, size_t len)
 {
 	(void)hal_strncpy(features, "Using Triple Timer Counter", len);
 	features[len - 1U] = '\0';
