+++ hal/armv8m/cpu.c
@@ -226,7 +226,7 @@ char *hal_cpuInfo(char *info)
 }
 
 
-char *hal_cpuFeatures(char *features, unsigned int len)
+char *hal_cpuFeatures(char *features, size_t len)
 {
 	unsigned int n = 0;
 #if KERNEL_FPU_SUPPORT
